bugfix
(1) When you select “New Column”, you get an error when the column name contains uppercase letters.
(2) Click “New Index for Selected Columns” for columns named with uppercase letters and get error.
(3) Clicking “New Constraints for Selected Columns” for a column named with uppercase letters gives an error.
(4) Error occurs when clicking “Read Data in SQL Console” for a column named with uppercase letters.
(5) Error occurs when a database object with a column named with uppercase letters is executed after copying the SELECT SQL.
(6) Error occurs when copying INSERT SQL after execution of database objects with columns named with uppercase alphabets.
(7) Database object with columns named with uppercase alphabets errors after copying UPDATE SQL.
(9) An error occurs when a database object with a column named with uppercase letters is executed after “Edit Data” and then “Save”.
(10) When a new column is created and the column name contains uppercase letters, it is automatically changed to lowercase.